Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-syntax/unclosed-constructs-expected.txt


PASS "[foo]" is a valid selector
PASS "[foo" is a valid selector
PASS ":nth-child(1)" is a valid selector
PASS ":nth-child(1" is a valid selector
